Move existing Nodes tests under Node namespace
